@@ -1,8 +1,8 @@
 package ru.app.churchofchrist;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.View;
 
 import com.kobakei.ratethisapp.RateThisApp;
